refactor(client): rename AuthLayout into CardLayout
