+++ packages/client/src/elements/layouts/CardLayout/CardLayout.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
 
-import { AuthLayoutProps } from './AuthLayout.types';
+import { CardLayoutProps } from './CardLayout.types';
 import { Container } from './index';
 
 export default {
@@ -9,7 +9,7 @@ export default {
   component: Container,
 } as Meta;
 
-const Template: Story<AuthLayoutProps> = (args) => <Container {...args} />;
+const Template: Story<CardLayoutProps> = (args) => <Container {...args} />;
 
 export const Default = Template.bind({});
 Default.args = {
