+++ packages/client/src/elements/layouts/index.ts
@@ -1,3 +1,3 @@
-export { default as AuthLayout } from './AuthLayout';
+export { default as CardLayout } from './CardLayout';
 export { default as DefaultLayout } from './DefaultLayout';
 export { default as GlobalLayout } from './GlobalLayout';
